增加了查看汇总的功能等

+++ public/js/admin.js
@@ -1,20 +1,43 @@
 $(function(){
+	$('.jj').click(function(e) {
+		var target = $(e.target)
+		var id = target.data('id')
+		var tr = $('.item-id-' + id);
+
+	})
 	$('.del').click(function(e) {
 		var target = $(e.target)
 		var id = target.data('id')
 		var tr = $('.item-id-' + id)
-
-		$.ajax({
-			type: 'DELETE',
-			url: '/admin/record/?id=' + id
-		})
-		.done(function(results){
-			if(results.success === 1){
-				if(tr.length > 0){
-					tr.remove()
-				}
-			}
-		})
+		bootbox.confirm({		
+	    title: "删除该记录？",
+	    message: "是否删除本条记录？",
+	    buttons: {
+        cancel: {
+          label: '取消',
+          className: 'btn-success'
+        },
+        confirm: {
+          label: '确认',
+          className: 'btn-danger'
+        }
+	    },
+	    callback: function (result) {
+	    	if(result){
+					$.ajax({
+						type: 'DELETE',
+						url: '/admin/record/?id=' + id
+					})
+					.done(function(results){
+						if(results.success === 1){
+							if(tr.length > 0){
+								tr.remove()
+							}
+						}
+					})
+	    	}
+	    }
+		});	
 	})
 
 	$('.pay').click(function(e) {
